+++ src/pe_testenvironment/TestParticleGUI.java
@@ -118,7 +118,7 @@ public class TestParticleGUI extends JWindow {
 		super.paint(g);
 		g.setColor(Color.green);
 		for (Particle p : pm.particlesystem) {
-			g.fillOval(p.getLocation(0), p.getLocation(1), 40, 40);
+			g.fillOval((int) p.getLocation(0), (int) p.getLocation(1), 40, 40);
 		}
 	}
